requirements.txt
pandas
SQLAlchemy
psycopg2-binary
apache-airflow-providers-postgres
